Add correction for esEnergyOverSCRawEnergy (preshower variable)
In UL2018, we saw that the correction workflow that worked for 16 and 17 (i.e. chained training+correction of 6 shower shapes variables + 3 isolations) doesn't work as expected.
To put a patch on this, we correct also the variable esEnergyOverSCRawEnergy (ratio between preshower energy and raw energy of the supercluster) using the corrected shower shapes
as additional input to the regressors.
The training was performed applying the additional cut abs(ScEta) > 1.653, which allows to treat the newly corrected variable like the 6 shower shapes (i.e., no peak-to-tail classifier is trained).

In this commit, we implement the new correction in the PhoIdInputsCorrector workflow. Here are the main steps:

- a new field called "preshower", with the specifications of the new variable to correct, is added to
/eos/cms/store/group/phys_higgs/cmshgg/flashgg-data/Taggers/data/PhoIdInputsCorrections/corrections_summary_2018_Legacy.json
- Taggers/python/flashggDifferentialPhoIdInputsCorrection_cfi.py and the associated plugin Taggers/plugins/DifferentialPhoIdInputsCorrector.cc get a new case other that "shower_shapes" and "isolations", called
"preshower"
- the way in which the preshower variable is fetched when the PhoID is computed is changed in MicroAOD/src/PhotonIdUtils.cc